refactor temperature formulae
