+++ src/UnitConverter/Calculator/Formula/Temperature/Celsius/ToFahrenheit.php
@@ -12,24 +12,24 @@ declare(strict_types = 1);
  * file that was distributed with this source code.
  */
 
-namespace UnitConverter\Calculator\Formula\Celsius;
+namespace UnitConverter\Calculator\Formula\Temperature\Celsius;
 
-use UnitConverter\Calculator\Formula\AbstractFormula;
+use UnitConverter\Calculator\Formula\Temperature\TemperatureFormula;
 
 /**
- * Formula to convert Celsius values to Kelvin.
+ * Formula to convert Celsius values to Fahrenheit.
  *
  * @version 1.0.0
  * @since 0.8.1
  * @author Jordan Brauer <jbrauer.inc@gmail.com>
  */
-class ToKelvin extends AbstractFormula
+class ToFahrenheit extends TemperatureFormula
 {
-    const FORMULA_STRING = 'K = °C + 273.15';
+    const FORMULA_STRING = '°F = (°C × (9 ÷ 5)) + 32';
 
-    const FORMULA_TEMPLATE = '%s K = %s°C + 273.15';
+    const FORMULA_TEMPLATE = '%s°F = (%s°C × (9 ÷ 5)) + 32';
 
-    const MAGIC_NUMBER = 273.15;
+    const MAGIC_NUMBER = 32;
 
     /**
      * {@inheritDoc}
@@ -37,7 +37,9 @@ class ToKelvin extends AbstractFormula
     public function describe($value, $fromUnits, $toUnits, int $precision = null)
     {
         // XXX: this formula assumes all calculators can accept strings, as it's the safest type.
-        $addResult = $this->calculator->add($value, (string) self::MAGIC_NUMBER);
+        $divisor = $this->calculator->div('9', '5');
+        $mulResult = $this->calculator->mul($value, $divisor);
+        $addResult = $this->calculator->add($mulResult, (string) self::MAGIC_NUMBER);
         $result = $this->calculator->round($addResult, $precision);
 
         $this->plugVariables($result, $value);
